requirements_test.txt
# Testing requirements for Drivee integration
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=4.1.0

# Integration requirements (from manifest.json)
aiohttp>=3.12.15
tenacity>=8.0.0
pydantic>=2.11.7
driveeClient==0.1.4
cachetools>=5.3.0
